explicit cast to float just to be safe for DogRenderer anim_timeline

+++ src/main/java/doggytalents/client/entity/render/DogRenderer.java
@@ -78,7 +78,7 @@ public class DogRenderer extends MobRenderer<Dog, DogModel> {
 
     @Override
     protected float getBob(Dog dog, float pticks) {
-        return dog.tickCount + pticks;
+        return ((float)dog.tickCount) + pticks;
     }
 
     @Override
